add ApplicationXMLParser project
ApplicationXMLParser to be integrated into FD's air packaging process.
it creates a copy of application.xml and filters out nodes by a condition attribute. the condition value is checked against the compiler constants found in the flex config generated by fdbuild. only if a constant is defined and equals "true" the node stays in the xml tree.